+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `--json` flag to `alias` command, [PR-287](https://github.com/reductstore/reduct-cli/pull/287) by @vbmade2000
 - Add `--json` flag to `replica` command, [PR-288](https://github.com/reductstore/reduct-cli/pull/288) by @vbmade2000
 - Add `--json` flag to `attachment` command, [PR-289](https://github.com/reductstore/reduct-cli/pull/289) by @vbmade2000
+- Add `--json` flag to `lifecycle` command, [PR-290](https://github.com/reductstore/reduct-cli/pull/290) by @vbmade2000
 
 ### Changed